feat/필터 검색 기능 추가

+++ src/main/java/daewoo/team5/hotelreservation/domain/hotel/dto/ReservationDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class ReservationDTO {
 
     private Long reservationId;
-    private int userId; // 변경: userId는 int 타입
+    private Long userId; // 변경: userId는 int 타입
     private String roomId;
     private String status;
     private BigDecimal amount;
